Only launch if variable is set to true
Launching involves rebuilding the app, which takes quite a while.
This introduces a variables in "defaults" which allows you to
set whether the launch is done.

This variable is set to false in the unict-openstack molecule scenario
to speed up CI.

In order to do this, I had to move a task from "launch.yml" to
"install-discourse.yml" which replaces the acceptable storage
drivers in `launcher.sh`